requirements.txt
camelot-py==0.9.0
cffi==1.15.1
chardet==5.0.0
charset-normalizer==3.0.1
click==8.1.3
cryptography==38.0.3
et-xmlfile==1.1.0
icalendar==5.0.2
numpy==1.23.4
opencv-python==4.6.0.66
openpyxl==3.0.10
pandas==1.5.1
pdfminer.six==20221105
pycparser==2.21
PyPDF2==2.11.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
